+++ assets/dags/mwaa_dr/backup_metadata.py
@@ -15,55 +15,16 @@ IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 """
 
-from airflow import version, DAG
+from airflow import DAG
 from airflow.models import Variable
+from mwaa_dr.dr_factory import DRFactory
+
 
 kwargs = {
     "dag_id": "backup_metadata",
     "path_prefix": "data",
     "storage_type": Variable.get("DR_STORAGE_TYPE", default_var="S3"),
 }
-airflow_version = version.version
 
-factory = None
-if airflow_version.startswith("2.4"):
-    from mwaa_dr.v_2_4.dr_factory import DRFactory_2_4
 
-    factory = DRFactory_2_4(**kwargs)
-
-elif airflow_version.startswith("2.5"):
-    from mwaa_dr.v_2_5.dr_factory import DRFactory_2_5
-
-    factory = DRFactory_2_5(**kwargs)
-
-elif airflow_version.startswith("2.6"):
-    from mwaa_dr.v_2_6.dr_factory import DRFactory_2_6
-
-    factory = DRFactory_2_6(**kwargs)
-
-elif airflow_version.startswith("2.7"):
-    from mwaa_dr.v_2_7.dr_factory import DRFactory_2_7
-
-    factory = DRFactory_2_7(**kwargs)
-
-elif airflow_version.startswith("2.8"):
-    from mwaa_dr.v_2_8.dr_factory import DRFactory_2_8
-
-    factory = DRFactory_2_8(**kwargs)
-
-elif airflow_version.startswith("2.9"):
-    from mwaa_dr.v_2_9.dr_factory import DRFactory_2_9
-
-    factory = DRFactory_2_9(**kwargs)
-
-elif airflow_version.startswith("2.10"):
-    from mwaa_dr.v_2_10.dr_factory import DRFactory_2_10
-
-    factory = DRFactory_2_10(**kwargs)
-
-else:
-    from mwaa_dr.framework.factory.default_dag_factory import DefaultDagFactory
-
-    factory = DefaultDagFactory(**kwargs)
-
-dag: DAG = factory.create_backup_dag()
+dag: DAG = DRFactory(**kwargs).create_backup_dag()